Refactor Application as application package
