+++ Server/online.quar/src/main/java/online/quar/application/Application.java
@@ -1,13 +1,13 @@
-package online.quar.security;
+package online.quar.application;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class SecuringWebApplication {
+public class Application {
 
     public static void main(String[] args) throws Throwable {
-        SpringApplication.run(SecuringWebApplication.class, args);
+        SpringApplication.run(Application.class, args);
     }
 
 }
