+++ Server/online.quar/src/main/java/online/quar/application/WebSecurityConfig.java
@@ -1,4 +1,4 @@
-package online.quar.security;
+package online.quar.application;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
